Move processResults implementation in ComparisonSlice
