+++ src/components/CompareResults/beta/ResultsTable.tsx
@@ -6,13 +6,9 @@ import TableContainer from '@mui/material/TableContainer';
 import { style } from 'typestyle';
 
 import { useAppSelector } from '../../../hooks/app';
-import { Strings } from '../../../resources/Strings';
+import { selectProcessedResults } from '../../../reducers/ComparisonSlice';
 import { Colors, Spacing } from '../../../styles';
-import type {
-  CompareResultsItem,
-  RevisionsHeader,
-  ThemeMode,
-} from '../../../types/state';
+import type { ThemeMode } from '../../../types/state';
 import NoResultsFound from './NoResultsFound';
 import TableContent from './TableContent';
 import TableHeader from './TableHeader';
@@ -21,71 +17,12 @@ const customStyles = {
   boxShadow: 'none',
 };
 
-type Results = {
-  key: string;
-  value: CompareResultsItem[];
-  revisionHeader: RevisionsHeader;
-};
-
-function processResults(results: CompareResultsItem[]) {
-  const processedResults: Map<string, CompareResultsItem[]> = new Map<
-    string,
-    CompareResultsItem[]
-  >();
-  results.forEach((result) => {
-    const { new_rev: newRevision, header_name: header } = result;
-    const rowIdentifier = header.concat(' ', newRevision);
-    if (processedResults.has(rowIdentifier)) {
-      (processedResults.get(rowIdentifier) as CompareResultsItem[]).push(
-        result,
-      );
-    } else {
-      processedResults.set(rowIdentifier, [result]);
-    }
-  });
-  const restructuredResults: Results[] = Array.from(
-    processedResults,
-    function (entry) {
-      return {
-        key: entry[0],
-        value: entry[1],
-        revisionHeader: {
-          suite: entry[1][0].suite,
-          framework_id: entry[1][0].framework_id,
-          test: entry[1][0].test,
-          option_name: entry[1][0].option_name,
-          extra_options: entry[1][0].extra_options,
-          new_rev: entry[1][0].new_rev,
-          new_repo: entry[1][0].new_repository_name,
-        },
-      };
-    },
-  );
-
-  return restructuredResults;
-}
-
 function ResultsTable(props: ResultsTableProps) {
   const { themeMode } = props;
-  const allRevisionsOption =
-    Strings.components.comparisonRevisionDropdown.allRevisions;
 
   const loading = useAppSelector((state) => state.compareResults.loading);
 
-  const processedResults = useAppSelector((state) => {
-    const { data } = state.compareResults;
-    const { activeComparison } = state.comparison;
-
-    if (activeComparison === allRevisionsOption) {
-      const allResults = ([] as CompareResultsItem[]).concat(
-        ...Object.values(data),
-      );
-      return processResults(allResults);
-    } else {
-      const results = data[activeComparison];
-      return processResults(results);
-    }
-  });
+  const processedResults = useAppSelector(selectProcessedResults);
 
   const themeColor100 =
     themeMode === 'light' ? Colors.Background100 : Colors.Background100Dark;
